Made public processResults method
For new HS attended Epic, we need to cover a case: we are supposed to keep the search results, even after selecting something.
The only way to proceed without touching the logics is to make the processResults method 'public' (accessible from a custom template).

The fix is working fine and do not cause any regression since we just making public a method.